+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- `services.addAuthWhitelistedIp()` resource implementation with tests and usage example
 - `services.deleteAuthWhitelistedIp()` resource implementation with tests and usage example
 - `services.requestProlongation()` resource implementation with tests and usage example
 - `services.getAuthWhitelistedIps()` resource implementation with tests and usage example